+++ apps/api-service/src/products/intelligence/intelligence.controller.ts
@@ -8,15 +8,15 @@ import {
   Param,
   BadRequestException,
 } from "@nestjs/common"
-import { OneAgentService } from "./oneagent.service"
-import { AIChatDto } from "./dto/ai-chat.dto"
+import { IntelligenceService } from "./intelligence.service"
+import { AIChatDto } from "./dto/chat.dto"
 import { AuthGuard, ModRequest } from "@/auth/auth.guard"
-import { AISummarizeDto } from "./dto/ai-summarize.dto"
+import { AISummarizeDto } from "./dto/summarize.dto"
 import { statusMessages } from "@/shared/constants/status-messages"
 
-@Controller("oneagent")
-export class OneAgentController {
-  constructor(private readonly service: OneAgentService) {}
+@Controller("intelligence")
+export class IntelligenceController {
+  constructor(private readonly service: IntelligenceService) {}
 
   @UseGuards(AuthGuard)
   @Get("thread/:threadId")