+++ apps/api-service/src/products/intelligence/intelligence.module.ts
@@ -1,15 +1,15 @@
 import { Module } from "@nestjs/common"
-import { OneAgentService } from "./oneagent.service"
-import { OneAgentController } from "./oneagent.controller"
+import { IntelligenceService } from "./intelligence.service"
+import { IntelligenceController } from "./intelligence.controller"
 import { CqrsModule } from "@nestjs/cqrs"
 import { ProductsDbConnectionMap } from "@/shared/constants/db-connection.map"
 import { Thread, ThreadSchema } from "./schemas/thread.schema"
 import { CreateThreadCommandHandler } from "./commands/handler/create-thread.handler"
-import { OneAgentRepository } from "./oneagent.repository"
+import { IntelligenceRepository } from "./intelligence.repository"
 import { EntityModule } from "@/shared/entity/entity.module"
 import { FetchThreadByIdQueryHandler } from "./queries/handler/fetch-thread-by-id.handler"
 import { HttpModule } from "@nestjs/axios"
-import { OneAgentStrategy } from "./oneagent.strategy"
+import { IntelligenceStrategy } from "./intelligence.strategy"
 import { ChatTools } from "./tools/chat.tool"
 import { config } from "@/config"
 import { SummarizeTools } from "./tools/summarize.tool"
@@ -20,22 +20,22 @@ import { SummarizeTools } from "./tools/summarize.tool"
     CqrsModule,
     EntityModule.forRoot(
       config.PRODUCTS_DATABASE_URI,
-      ProductsDbConnectionMap.OneAgent
+      ProductsDbConnectionMap.Intelligence
     ),
     EntityModule.forFeature(
       [{ name: Thread.name, schema: ThreadSchema }],
-      ProductsDbConnectionMap.OneAgent
+      ProductsDbConnectionMap.Intelligence
     ),
   ],
-  controllers: [OneAgentController],
+  controllers: [IntelligenceController],
   providers: [
-    OneAgentService,
-    OneAgentRepository,
-    OneAgentStrategy,
+    IntelligenceService,
+    IntelligenceRepository,
+    IntelligenceStrategy,
     ChatTools,
     SummarizeTools,
     CreateThreadCommandHandler,
     FetchThreadByIdQueryHandler,
   ],
 })
-export class OneAgentModule {}
+export class IntelligenceModule {}