+++ apps/api-service/src/products/products.module.ts
@@ -1,5 +1,5 @@
 import { Module } from "@nestjs/common"
-import { OneAgentModule } from "./oneagent/oneagent.module"
+import { IntelligenceModule } from "./intelligence/intelligence.module"
 import { WealthAnalyzerModule } from "./wealthanalyzer/wealthanalyzer.module"
 import { DebtTrackModule } from "./debttrack/debttrack.module"
 import { TaxAdvisorModule } from "./taxadvisor/taxadvisor.module"
@@ -11,7 +11,7 @@ import { FinNewsModule } from "./finnews/finnews.module"
     TaxAdvisorModule,
     WealthAnalyzerModule,
     DebtTrackModule,
-    OneAgentModule,
+    IntelligenceModule,
     WealthGoalModule,
     FinNewsModule,
   ],